requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
asyncio

# API clients
tweepy>=4.14.0
PyGithub>=1.59.0
praw>=7.7.0
notion-client>=2.2.0
anthropic>=0.7.0

# Data processing
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Logging and monitoring
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0